remove production things

+++ publish.sh
@@ -1,26 +1,39 @@
 #!/bin/sh
 
-root_dir="$HOME/org/website"
+root_dir=$(pwd)
 build_dir=$root_dir/build
 site_dir=$root_dir/site
 dist_dir=$root_dir/dist
 
-rm -rf "$build_dir"
+. "$root_dir"/.env
+
+rm -rf "${build_dir:?}"
 mkdir "$build_dir"
-rm -rf "$dist_dir"
+rm -rf "${dist_dir:?}"
 mkdir "$dist_dir"
 
-cp -R "$site_dir"/* "$build_dir"
+cp -R "${site_dir:?}"/* "$build_dir"
 
 yarn parcel build "$build_dir/*.html" --no-source-maps --no-cache
 
 # Static content
-rsync --delete -av -e ssh ~/org/website/dist/ dingerdonger02_nothingsinside@ssh.nyc1.nearlyfreespeech.net:/home/public/nothingsinside.org/
+rsync -av -e ssh --delete "$root_dir"/dist/ "${static_dest:?}"
 
 # Scripts
-rsync -av -e ssh --delete ~/org/website/site/scripts/ dingerdonger02_nothingsinside@ssh.nyc1.nearlyfreespeech.net:/home/public/nothingsinside.org/scripts/
+rsync -av -e ssh --delete "$root_dir"/site/scripts/ "${scripts_dest:?}"
 
-# Protected
-rsync -av -e ssh ~/org/website/site/protected/ dingerdonger02_nothingsinside@ssh.nyc1.nearlyfreespeech.net:/home/protected/
+# Protected directory, for non user-facing data
+rsync -av -e ssh "$root_dir"/site/protected/
+
+
+#while getopts ":h" option; do
+#   case $option in
+#      h) # display Help
+#         Help
+#         exit;;
+#      *)
+#        exit;;
+#   esac
+#done
 
 #ssh dingerdonger02_nothingsinside@ssh.nyc1.nearlyfreespeech.net "chmod 666 /home/public/nothingsinside.org/scripts/visitor_count.txt"
